syslogd: implement checksum chains for log entries
This is a mechanism for detecting intrusion into log files.

Each time the server starts, it starts creating a checksum chain for
each log file. On restart, the chain starts with an empty checksum.

A mismatch between the checkum and the server restart is a possible
intrusion and the reason for investigation.

Each entry is accompanied by a checksum of this entry with the previous
checksum. So, to check the checksum of the current record, you need to
have a previous checksum.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>