syslogd: Drop message concatenation
Long messages are truncated and the remainder is appended to the next
message if possible.

Now the message parts will not be appended with the next one. These
parts will always be logged as separate messages.

This allows you to get rid of unnecessary copying of the message into a
temporary buffer and dynamic memory allocation during message
processing.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>